refactor: 딜러 생성자의 이름 파라미터 제거
- 딜러의 이름은 고정이기 때문에 이름 파라미터를 제거하였다.
