+++ src/test/java/blackjack/domain/gambler/DealerTest.java
@@ -1,6 +1,5 @@
 package blackjack.domain.gambler;
 
-import static blackjack.domain.Round.DEALER_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
@@ -15,7 +14,7 @@ public class DealerTest {
     @Test
     void getInitialCardsTest() {
         // given
-        Dealer dealer = new Dealer(DEALER_NAME);
+        Dealer dealer = new Dealer();
         Card card1 = new Card(CardShape.CLOVER, CardType.TEN);
         Card card2 = new Card(CardShape.HEART, CardType.EIGHT);
         dealer.addCard(card1);
